+++ packages/layer/src/styledContentContainer.jsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export default styled.div`
+  [data-role="droplistContent"] {
+    ${({ maxHeight }) => (
+      maxHeight ? `max-height: ${maxHeight}px` : ''
+    )}
+  }
+`;