+++ packages/media-card/src/links/card/styled.ts
@@ -1,9 +1,0 @@
-/* tslint:disable:variable-name */
-import styled from 'styled-components';
-import { Href, HrefProps } from '../../utils/href';
-
-export const A = styled(Href)`
-  color: initial;
-  // We need to do this to make TS happy
-  ${(props: HrefProps) => ''}
-`;