+++ packages/task-decision/stories/DecisionItem-story.tsx
@@ -1,6 +1,6 @@
 import { storiesOf } from '@kadira/storybook';
 import * as React from 'react';
-import { ReactRenderer as Renderer } from '@atlaskit/editor-core/dist/es5/renderer';
+import { ReactRenderer as Renderer } from '@atlaskit/renderer';
 
 import DecisionItem from '../src/components/DecisionItem';
 import { document, getParticipants } from '../src/support/story-data';